Add BluetoothProvider to CentralServicesViewController
