+++ ExampleApp/ExampleApp/Utilities/Error+Printable.swift
@@ -1,0 +1,13 @@
+import RxBluetoothKit
+
+extension Error {
+
+    var printable: String {
+        if let bleError = self as? BluetoothError {
+            return bleError.description
+        }
+
+        return localizedDescription
+    }
+
+}
